Add NoopLocation as standin for no Location from Driver
We want a means where a NApp can leave in isolation if so desired
by the user without the panic we added before, hence adding a new
NoopLocation which will be created to standard in for an NApp
Location field which then allows us use the app without having a
driver for cases where we wish to use it in such mode. Also, we
have changed the path to set either the URL.Path or URL.hash instead
of the URL.String or URL.hash when setting PushEvents.

Please enter the commit message for your changes. Lines starting